% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignum.R
\docType{package}
\name{lignum-package}
\alias{lignum}
\alias{lignum-package}
\title{lignum: Fast Decision Tree Inference Library}
\description{
A high-performance inference library for decision-tree based models, specifically optimized for 'LightGBM' and 'XGBoost'. By implementing a native C++ engine, the package bypasses the performance bottlenecks of standard parsers. It features instant file loading via memory mapping ('mmap'), zero-overhead architecture, and efficient row-major multi-threaded computations using OpenMP to ensure maximum CPU cache efficiency during batch predictions.
}
\details{
Lignum delivers fast inference for decision-tree based models.
}
\author{
\strong{Maintainer}: Kacper Piotrowski \email{misterekkk@users.noreply.github.com}

}
