% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignum.R
\docType{package}
\name{lignum}
\alias{lignum-package}
\alias{lignum}
\title{lignum: Fast Decision Tree Inference library}
\description{
High-performance inference library
}
\details{
Lignum deliveres fast interference for decision-tree based models.
}
\author{
\strong{Maintainer}: Kacper Piotrowski \email{tobedone@example.com}

}
