% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignum.R
\name{load_model}
\alias{load_model}
\title{Load a Lignum Model}
\usage{
load_model(filepath, format = "auto")
}
\arguments{
\item{filepath}{A string specifying the path to the model file.}

\item{format}{A string specifying the format of the model file.
Options are \code{'auto'}, \code{'lignum'}, \code{'binary'}, \code{'lightgbm'}, \code{'xgboost'}.
If \code{'auto'} (the default), the function attempts to load binary file.}
}
\value{
An object of class \code{lignum_model} ready for inference.
}
\description{
Load a trained model from a JSON or Lignum binary file.
}
\examples{
\dontrun{
model <- load_model("model_xgboost.json", format = "xgboost")
}
}
