% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignum.R
\name{save_model}
\alias{save_model}
\title{Save Lignum Model}
\usage{
save_model(object, filepath)
}
\arguments{
\item{object}{A \code{lignum_model} object.}

\item{filepath}{A string specifying the destination path for the \code{.bin} file.}
}
\description{
Saves the compiled model to an binary format for mmap-based loading.
}
\details{
Save the model to Lignum binary format.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
save_model(model, "fast_model.bin")
\dontshow{\}) # examplesIf}
}
